+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added 
+
+- Screens arrangement.
+
+## [1.0.0] - 2021-02-03
+
 ### Added
 
 - Rules Component.
 - Logo Component.
+- Score table Component.
+- Scores Component.
 - Reference Component.
 - Label Component.
 - Intro Component.
@@ -22,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Svg logo asset.
 - Types declaration for Typescript in `main.d.ts` file.
 - Github actions.
+
 
 ### Changed
 
@@ -62,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sass style loaders with autoprefixer for Webpack.
 - Stylelint for Sass.
 
-[unreleased]: https://github.com/ruljin/CodersCamp2020.Project.TypeScript.YahtzeeGame/compare/0.2.0...HEAD
+[unreleased]: https://github.com/ruljin/CodersCamp2020.Project.TypeScript.YahtzeeGame/compare/1.0.0...HEAD
+[1.0.0]: https://github.com/ruljin/CodersCamp2020.Project.TypeScript.YahtzeeGame/compare/0.2.0...1.0.0
 [0.2.0]: https://github.com/ruljin/CodersCamp2020.Project.TypeScript.YahtzeeGame/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/ruljin/CodersCamp2020.Project.TypeScript.YahtzeeGame/releases/tag/0.1.0